its working now

+++ OOPEksamensOpgave/Interfaces/ITallySystem.cs
@@ -11,7 +11,7 @@ namespace OOPEksamensOpgave.Interfaces
     public interface ITallySystem
     {
         IEnumerable<Product> ActiveProducts();
-        InsertCashTransaction AddCreditsToAccount(User user, int amount);
+        InsertCashTransaction AddCreditsToAccount(User user, decimal amount);
         BuyTransaction BuyProduct(User user, Product product);
         Product GetProductByID(int id);
         IEnumerable<Transaction> GetTransactions(User user, int count);
